Bug #5479 : JCR copy and move problems - adjusting some code after cherry pick of the commit from core-5.14.x - fixing the case of the copy that did not work too (history of versions was not respected, i18n was not supported, date of a version was not the one of the original but the date of the copy action !) - fixing the case of the move - upgrading the simple document access controller in order to take in account the versions of attachments - improving the mechanism of universal link associated to attachments - handling the language content for versioned attachments
Conflicts:
	lib-core/src/main/java/org/silverpeas/attachment/model/HistorisedDocument.java
	lib-core/src/main/java/org/silverpeas/attachment/model/SimpleDocument.java
	lib-core/src/main/java/org/silverpeas/attachment/repository/DocumentConverter.java
	lib-core/src/main/java/org/silverpeas/attachment/repository/DocumentRepository.java
	lib-core/src/test/java/org/silverpeas/attachment/repository/HistorisedDocumentRepositoryTest.java
	war-core/src/main/java/org/silverpeas/attachment/web/VersioningRequestRouter.java
	war-core/src/main/java/org/silverpeas/attachment/web/VersioningSessionController.java
	war-core/src/main/webapp/attachment/jsp/publicVersions.jsp
	web-core/src/main/java/org/silverpeas/accesscontrol/SimpleDocumentAccessController.java

Conflicts:
	lib-core/src/main/java/org/silverpeas/attachment/repository/DocumentConverter.java
	lib-core/src/main/java/org/silverpeas/attachment/repository/DocumentRepository.java

+++ lib-core/src/main/java/org/silverpeas/attachment/repository/DocumentRepository.java
@@ -149,17 +149,23 @@ public class DocumentRepository {
     targetDoc.setDocumentType(document.getDocumentType());
     targetDoc.setNodeName(document.getNodeName());
     prepareComponentAttachments(session, destination.getInstanceId(), document.getFolder());
+    session.save();
     Node originDocumentNode = session.getNodeByIdentifier(document.getPk().getId());
-    if (converter.isVersioned(originDocumentNode) && !originDocumentNode.isCheckedOut()) {
-      checkoutNode(originDocumentNode, document.getUpdatedBy());
-    }
     if (!originDocumentNode.getPath().equals(targetDoc.getFullJcrPath())) {
-      session.move(originDocumentNode.getPath(), targetDoc.getFullJcrPath());
+      session.getWorkspace().move(originDocumentNode.getPath(), targetDoc.getFullJcrPath());
     }
+    VersionManager versionManager = session.getWorkspace().getVersionManager();
     Node targetDocumentNode = session.getNode(targetDoc.getFullJcrPath());
+    if (converter.isVersionedMaster(targetDocumentNode) && !targetDocumentNode.isCheckedOut()) {
+      versionManager.checkout(targetDocumentNode.getPath());
+    }
     converter.addStringProperty(targetDocumentNode, SLV_PROPERTY_FOREIGN_KEY, destination.getId());
     converter.addStringProperty(targetDocumentNode, SLV_PROPERTY_INSTANCEID, destination.
         getInstanceId());
+    if (converter.isVersionedMaster(targetDocumentNode) && targetDocumentNode.isCheckedOut()) {
+      session.save();
+      versionManager.checkin(targetDocumentNode.getPath());
+    }
     pk.setId(targetDocumentNode.getIdentifier());
     return pk;
   }
@@ -211,24 +217,56 @@ public class DocumentRepository {
       WAPrimaryKey destination) throws RepositoryException, IOException {
     prepareComponentAttachments(destination.getInstanceId(), document.getFolder());
     SimpleDocumentPK pk = new SimpleDocumentPK(null, destination.getInstanceId());
-    List<SimpleDocument> history = document.getHistory();
+    List<SimpleDocument> history = new ArrayList<SimpleDocument>(document.getHistory());
+    Collections.reverse(history);
     history.add(document);
-    Collections.reverseOrder();
     SimpleDocument targetDoc = new HistorisedDocument(history.remove(0));
     targetDoc.setNodeName(null);
     targetDoc.setPK(pk);
     targetDoc.setDocumentType(document.getDocumentType());
     targetDoc.setForeignId(destination.getId());
+    targetDoc.setUpdatedBy(null);
     targetDoc.computeNodeName();
     pk = createDocument(session, targetDoc);
     unlock(session, targetDoc, false);
+    VersionManager versionManager = session.getWorkspace().getVersionManager();
+    String currentVersion = targetDoc.getVersion();
     for (SimpleDocument doc : history) {
-      lock(session, targetDoc, document.getUpdatedBy());
-      targetDoc = new HistorisedDocument(doc);
-      targetDoc.setPK(pk);
-      targetDoc.setForeignId(destination.getId());
-      updateDocument(session, targetDoc);
-      unlock(session, targetDoc, false);
+      if (I18NHelper.isI18nActivated()) {
+        // One language content is aimed by a version. So the first step here is to search the
+        // language content updated.
+        Set<String> checkedLanguages = new HashSet<String>();
+        checkedLanguages.add(doc.getLanguage());
+        for (String language : I18NHelper.getAllSupportedLanguages()) {
+          if (!checkedLanguages.contains(language)) {
+            SimpleDocument temp = findDocumentById(session, doc.getPk(), language);
+            if (temp != null && !checkedLanguages.contains(temp.getLanguage()) &&
+                temp.getUpdated().after(doc.getUpdated())) {
+              doc = temp;
+            }
+            checkedLanguages.add(language);
+          }
+        }
+      }
+      HistorisedDocument targetHistorisedDoc = new HistorisedDocument(doc);
+      targetHistorisedDoc.setPK(pk);
+      targetHistorisedDoc.setForeignId(destination.getId());
+      targetHistorisedDoc.setNodeName(targetDoc.getNodeName());
+      Node masterDocumentNode = session.getNodeByIdentifier(pk.getId());
+      if (!currentVersion.equals(doc.getVersion())) {
+        // In this case, a functional version is performed, so the common tools are used
+        lock(session, targetDoc, doc.getUpdatedBy());
+        converter.fillNode(targetHistorisedDoc, masterDocumentNode);
+        unlock(session, targetHistorisedDoc, false);
+        currentVersion = targetHistorisedDoc.getVersion();
+      } else {
+        // In this case, the version is a technical one (setOrder for example) and technical
+        // tools are used
+        versionManager.checkout(masterDocumentNode.getPath());
+        converter.fillNode(targetHistorisedDoc, masterDocumentNode);
+        session.save();
+        versionManager.checkin(masterDocumentNode.getPath());
+      }
     }
     return pk;
   }
@@ -332,8 +370,8 @@ public class DocumentRepository {
       String comment) throws RepositoryException, IOException {
     try {
       Node documentNode = session.getNodeByIdentifier(documentPk.getId());
-      boolean versionedNode = documentNode.getParent() instanceof Version || converter.isVersioned(
-          documentNode);
+      boolean versionedNode =
+          documentNode.getParent() instanceof Version || converter.isVersionedMaster(documentNode);
       Node parent = documentNode.getParent();
       if (parent instanceof Version) {
         Version selectedVersion = (Version) parent;
@@ -385,7 +423,7 @@ public class DocumentRepository {
 
   private void deleteDocumentNode(Node documentNode) throws RepositoryException {
     if (null != documentNode) {
-      if (converter.isVersioned(documentNode)) {
+      if (converter.isVersionedMaster(documentNode)) {
         removeHistory(documentNode);
       }
       documentNode.remove();
@@ -628,9 +666,11 @@ public class DocumentRepository {
     Selector source = factory.selector(SLV_SIMPLE_DOCUMENT, SIMPLE_DOCUMENT_ALIAS);
     DescendantNode descendantNodeConstraint = factory.descendantNode(SIMPLE_DOCUMENT_ALIAS, session.
         getRootNode().getPath() + instanceId + '/');
-    Comparison foreignIdComparison = factory.comparison(factory.propertyValue(SIMPLE_DOCUMENT_ALIAS,
-        SLV_PROPERTY_FOREIGN_KEY), QueryObjectModelFactory.JCR_OPERATOR_EQUAL_TO, factory.
-        literal(session.getValueFactory().createValue(foreignId)));
+    Comparison foreignIdComparison = factory
+        .comparison(factory.propertyValue(SIMPLE_DOCUMENT_ALIAS, SLV_PROPERTY_FOREIGN_KEY),
+            QueryObjectModelFactory.JCR_OPERATOR_EQUAL_TO, factory.
+                literal(session.getValueFactory().createValue(foreignId))
+        );
     Ordering order = factory.ascending(factory.propertyValue(SIMPLE_DOCUMENT_ALIAS,
         SLV_PROPERTY_ORDER));
     QueryObjectModel query = factory.createQuery(source, factory.and(descendantNodeConstraint,
@@ -674,8 +714,7 @@ public class DocumentRepository {
    */
   public List<SimpleDocument> listExpiringDocuments(Session session, Date expiryDate,
       String language) throws RepositoryException {
-    NodeIterator iter = selectExpiringDocuments(session, DateUtil.getBeginOfDay(
-        expiryDate));
+    NodeIterator iter = selectExpiringDocuments(session, DateUtil.getBeginOfDay(expiryDate));
     return converter.convertNodeIterator(iter, language);
   }
 
@@ -690,8 +729,7 @@ public class DocumentRepository {
    */
   public List<SimpleDocument> listDocumentsRequiringWarning(Session session, Date alertDate,
       String language) throws RepositoryException {
-    NodeIterator iter = selectWarningDocuments(session, DateUtil.getBeginOfDay(
-        alertDate));
+    NodeIterator iter = selectWarningDocuments(session, DateUtil.getBeginOfDay(alertDate));
     return converter.convertNodeIterator(iter, language);
   }
 
@@ -848,7 +886,7 @@ public class DocumentRepository {
   public void addContent(Session session, SimpleDocumentPK documentPk, SimpleAttachment attachment)
       throws RepositoryException {
     Node documentNode = session.getNodeByIdentifier(documentPk.getId());
-    if (converter.isVersioned(documentNode) && !documentNode.isCheckedOut()) {
+    if (converter.isVersionedMaster(documentNode) && !documentNode.isCheckedOut()) {
       String owner = attachment.getUpdatedBy();
       if (!StringUtil.isDefined(owner)) {
         owner = attachment.getCreatedBy();
@@ -893,7 +931,7 @@ public class DocumentRepository {
   public boolean removeContent(Session session, SimpleDocumentPK documentPk, String language)
       throws RepositoryException {
     Node documentNode = session.getNodeByIdentifier(documentPk.getId());
-    if (converter.isVersioned(documentNode) && !documentNode.isCheckedOut()) {
+    if (converter.isVersionedMaster(documentNode) && !documentNode.isCheckedOut()) {
       checkoutNode(documentNode, null);
     }
     converter.removeAttachment(documentNode, language);
@@ -1041,7 +1079,7 @@ public class DocumentRepository {
   public void setVersionnable(Session session, SimpleDocumentPK documentPk) throws
       RepositoryException {
     Node documentNode = session.getNodeByIdentifier(documentPk.getId());
-    if (!converter.isVersioned(documentNode)) {
+    if (!converter.isVersionedMaster(documentNode)) {
       documentNode.addMixin(MIX_SIMPLE_VERSIONABLE);
       documentNode.setProperty(SLV_PROPERTY_VERSIONED, true);
     }
@@ -1058,7 +1096,7 @@ public class DocumentRepository {
   public void removeVersionnable(Session session, SimpleDocumentPK documentPk) throws
       RepositoryException {
     Node documentNode = session.getNodeByIdentifier(documentPk.getId());
-    if (converter.isVersioned(documentNode)) {
+    if (converter.isVersionedMaster(documentNode)) {
       removeHistory(documentNode);
       VersionHistory history = documentNode.getSession().getWorkspace().getVersionManager().
           getVersionHistory(documentNode.getPath());
@@ -1145,12 +1183,12 @@ public class DocumentRepository {
     String targetDir = copy.getDirectoryPath(null);
     targetDir = targetDir.replace('/', File.separatorChar);
     File target = new File(targetDir).getParentFile();
-    if (target.exists()) {
-      FileUtils.cleanDirectory(target);
-    }
     File source = new File(originDir).getParentFile();
     if (!source.exists() || !source.isDirectory() || source.listFiles() == null) {
       return;
+    }
+    if (target.exists()) {
+      FileUtils.cleanDirectory(target);
     }
     FileUtils.copyDirectory(source, target);
   }
@@ -1163,6 +1201,9 @@ public class DocumentRepository {
     File source = new File(originDir).getParentFile().getParentFile();
     if (!source.exists() || !source.isDirectory() || source.listFiles() == null) {
       return;
+    }
+    if (target.exists()) {
+      FileUtils.cleanDirectory(target);
     }
     FileUtils.copyDirectory(source, target);
   }
@@ -1179,6 +1220,21 @@ public class DocumentRepository {
     if (!target.getParentFile().getName().equals(source.getParentFile().getName())) {
       source = source.getParentFile();
       target = target.getParentFile();
+    }
+    if (!source.equals(target)) {
+      FileUtils.moveDirectory(source, target);
+      FileUtil.deleteEmptyDir(source.getParentFile());
+    }
+  }
+
+  public void moveFullContent(SimpleDocument origin, SimpleDocument copy) throws IOException {
+    String originDir = origin.getDirectoryPath(null);
+    File source = new File(originDir).getParentFile().getParentFile();
+    String targetDir = copy.getDirectoryPath(null);
+    targetDir = targetDir.replace('/', File.separatorChar);
+    File target = new File(targetDir).getParentFile().getParentFile();
+    if (!source.exists() || !source.isDirectory() || source.listFiles() == null) {
+      return;
     }
     if (!source.equals(target)) {
       FileUtils.moveDirectory(source, target);
